+++ sea-orm-codegen/src/entity/column.rs
@@ -1,3 +1,4 @@
+use crate::util::escape_rust_keyword;
 use heck::{CamelCase, SnakeCase};
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote};
@@ -14,11 +15,11 @@ pub struct Column {
 
 impl Column {
     pub fn get_name_snake_case(&self) -> Ident {
-        format_ident!("{}", self.name.to_snake_case())
+        format_ident!("{}", escape_rust_keyword(self.name.to_snake_case()))
     }
 
     pub fn get_name_camel_case(&self) -> Ident {
-        format_ident!("{}", self.name.to_camel_case())
+        format_ident!("{}", escape_rust_keyword(self.name.to_camel_case()))
     }
 
     pub fn get_rs_type(&self) -> TokenStream {
@@ -27,8 +28,6 @@ impl Column {
             ColumnType::Char(_)
             | ColumnType::String(_)
             | ColumnType::Text
-            | ColumnType::Time(_)
-            | ColumnType::Date
             | ColumnType::Custom(_) => "String",
             ColumnType::TinyInteger(_) => "i8",
             ColumnType::SmallInteger(_) => "i16",
@@ -37,6 +36,8 @@ impl Column {
             ColumnType::Float(_) => "f32",
             ColumnType::Double(_) => "f64",
             ColumnType::Json | ColumnType::JsonBinary => "Json",
+            ColumnType::Date => "Date",
+            ColumnType::Time(_) => "Time",
             ColumnType::DateTime(_) | ColumnType::Timestamp(_) => "DateTime",
             ColumnType::TimestampWithTimeZone(_) => "DateTimeWithTimeZone",
             ColumnType::Decimal(_) | ColumnType::Money(_) => "Decimal",
@@ -194,6 +195,11 @@ mod tests {
             make_col!("CAKE_FILLING_ID", ColumnType::Double(None)),
             make_col!("CAKE-FILLING-ID", ColumnType::Binary(None)),
             make_col!("CAKE", ColumnType::Boolean),
+            make_col!("date", ColumnType::Date),
+            make_col!("time", ColumnType::Time(None)),
+            make_col!("date_time", ColumnType::DateTime(None)),
+            make_col!("timestamp", ColumnType::Timestamp(None)),
+            make_col!("timestamp_tz", ColumnType::TimestampWithTimeZone(None)),
         ]
     }
 
@@ -211,6 +217,11 @@ mod tests {
             "cake_filling_id",
             "cake_filling_id",
             "cake",
+            "date",
+            "time",
+            "date_time",
+            "timestamp",
+            "timestamp_tz",
         ];
         for (col, snack_case) in columns.into_iter().zip(snack_cases) {
             assert_eq!(col.get_name_snake_case().to_string(), snack_case);
@@ -231,6 +242,11 @@ mod tests {
             "CakeFillingId",
             "CakeFillingId",
             "Cake",
+            "Date",
+            "Time",
+            "DateTime",
+            "Timestamp",
+            "TimestampTz",
         ];
         for (col, camel_case) in columns.into_iter().zip(camel_cases) {
             assert_eq!(col.get_name_camel_case().to_string(), camel_case);
@@ -241,7 +257,21 @@ mod tests {
     fn test_get_rs_type() {
         let columns = setup();
         let rs_types = vec![
-            "String", "String", "i8", "i16", "i32", "i64", "f32", "f64", "Vec<u8>", "bool",
+            "String",
+            "String",
+            "i8",
+            "i16",
+            "i32",
+            "i64",
+            "f32",
+            "f64",
+            "Vec<u8>",
+            "bool",
+            "Date",
+            "Time",
+            "DateTime",
+            "DateTime",
+            "DateTimeWithTimeZone",
         ];
         for (mut col, rs_type) in columns.into_iter().zip(rs_types) {
             let rs_type: TokenStream = rs_type.parse().unwrap();
@@ -271,6 +301,11 @@ mod tests {
             "ColumnType::Double.def()",
             "ColumnType::Binary.def()",
             "ColumnType::Boolean.def()",
+            "ColumnType::Date.def()",
+            "ColumnType::Time.def()",
+            "ColumnType::DateTime.def()",
+            "ColumnType::Timestamp.def()",
+            "ColumnType::TimestampWithTimeZone.def()",
         ];
         for (mut col, col_def) in columns.into_iter().zip(col_defs) {
             let mut col_def: TokenStream = col_def.parse().unwrap();