+++ sea-orm-codegen/src/util.rs
@@ -1,0 +1,23 @@
+pub(crate) fn escape_rust_keyword<T>(string: T) -> String
+where
+    T: ToString,
+{
+    let string = string.to_string();
+    if RUST_KEYWORDS.iter().any(|s| s.eq(&string)) {
+        format!("r#{}", string)
+    } else if RUST_SPECIAL_KEYWORDS.iter().any(|s| s.eq(&string)) {
+        format!("{}_", string)
+    } else {
+        string
+    }
+}
+
+pub(crate) const RUST_KEYWORDS: [&str; 49] = [
+    "as", "async", "await", "break", "const", "continue", "dyn", "else", "enum", "extern", "false",
+    "fn", "for", "if", "impl", "in", "let", "loop", "match", "mod", "move", "mut", "pub", "ref",
+    "return", "static", "struct", "super", "trait", "true", "type", "union", "unsafe", "use",
+    "where", "while", "abstract", "become", "box", "do", "final", "macro", "override", "priv",
+    "try", "typeof", "unsized", "virtual", "yield",
+];
+
+pub(crate) const RUST_SPECIAL_KEYWORDS: [&str; 3] = ["crate", "Self", "self"];