+++ src/entity/active_model.rs
@@ -1,8 +1,8 @@
 use crate::{
-    error::*, DatabaseConnection, DeleteResult, EntityTrait, Iterable, PrimaryKeyToColumn,
-    PrimaryKeyTrait, Value,
+    error::*, ConnectionTrait, DeleteResult, EntityTrait, Iterable, PrimaryKeyToColumn, Value,
 };
 use async_trait::async_trait;
+use sea_query::ValueTuple;
 use std::fmt::Debug;
 
 #[derive(Clone, Debug, Default)]
@@ -10,7 +10,8 @@ pub struct ActiveValue<V>
 where
     V: Into<Value>,
 {
-    value: Option<V>,
+    // Don't want to call ActiveValue::unwrap() and cause panic
+    pub(self) value: Option<V>,
     state: ActiveValueState,
 }
 
@@ -67,33 +68,64 @@ pub trait ActiveModelTrait: Clone + Debug {
 
     fn default() -> Self;
 
-    async fn insert(self, db: &DatabaseConnection) -> Result<Self, DbErr>
-    where
-        Self: ActiveModelBehavior,
-        <Self::Entity as EntityTrait>::Model: IntoActiveModel<Self>,
-    {
-        let am = ActiveModelBehavior::before_save(self, true, db)?;
-        let res = <Self::Entity as EntityTrait>::insert(am).exec(db).await?;
-        // Assume valid last_insert_id is not equals to Default::default()
-        if res.last_insert_id
-            != <<Self::Entity as EntityTrait>::PrimaryKey as PrimaryKeyTrait>::ValueType::default()
-        {
-            let found = <Self::Entity as EntityTrait>::find_by_id(res.last_insert_id)
-                .one(db)
-                .await?;
-            let am = match found {
-                Some(model) => Ok(model.into_active_model()),
-                None => Err(DbErr::Exec("Failed to find inserted item".to_owned())),
-            }?;
-            ActiveModelBehavior::after_save(am, true, db)
-        } else {
-            Ok(Self::default())
+    #[allow(clippy::question_mark)]
+    fn get_primary_key_value(&self) -> Option<ValueTuple> {
+        let mut cols = <Self::Entity as EntityTrait>::PrimaryKey::iter();
+        macro_rules! next {
+            () => {
+                if let Some(col) = cols.next() {
+                    if let Some(val) = self.get(col.into_column()).value {
+                        val
+                    } else {
+                        return None;
+                    }
+                } else {
+                    return None;
+                }
+            };
+        }
+        match <Self::Entity as EntityTrait>::PrimaryKey::iter().count() {
+            1 => {
+                let s1 = next!();
+                Some(ValueTuple::One(s1))
+            }
+            2 => {
+                let s1 = next!();
+                let s2 = next!();
+                Some(ValueTuple::Two(s1, s2))
+            }
+            3 => {
+                let s1 = next!();
+                let s2 = next!();
+                let s3 = next!();
+                Some(ValueTuple::Three(s1, s2, s3))
+            }
+            _ => panic!("The arity cannot be larger than 3"),
         }
     }
 
-    async fn update(self, db: &DatabaseConnection) -> Result<Self, DbErr>
+    async fn insert<'a, C>(self, db: &'a C) -> Result<Self, DbErr>
     where
         Self: ActiveModelBehavior,
+        <Self::Entity as EntityTrait>::Model: IntoActiveModel<Self>,
+        C: ConnectionTrait<'a>,
+        Self: 'a,
+    {
+        let am = ActiveModelBehavior::before_save(self, true, db)?;
+        let res = <Self::Entity as EntityTrait>::insert(am).exec(db).await?;
+        let found = <Self::Entity as EntityTrait>::find_by_id(res.last_insert_id)
+            .one(db)
+            .await?;
+        match found {
+            Some(model) => Ok(model.into_active_model()),
+            None => Err(DbErr::Exec("Failed to find inserted item".to_owned())),
+        }
+    }
+
+    async fn update<'a, C>(self, db: &'a C) -> Result<Self, DbErr>
+    where
+        C: ConnectionTrait<'a>,
+        Self: 'a,
     {
         let am = ActiveModelBehavior::before_save(self, false, db)?;
         let am = Self::Entity::update(am).exec(db).await?;
@@ -102,10 +134,11 @@ pub trait ActiveModelTrait: Clone + Debug {
 
     /// Insert the model if primary key is unset, update otherwise.
     /// Only works if the entity has auto increment primary key.
-    async fn save(self, db: &DatabaseConnection) -> Result<Self, DbErr>
+    async fn save<'a, C>(self, db: &'a C) -> Result<Self, DbErr>
     where
-        Self: ActiveModelBehavior,
+        Self: ActiveModelBehavior + 'a,
         <Self::Entity as EntityTrait>::Model: IntoActiveModel<Self>,
+        C: ConnectionTrait<'a>,
     {
         let mut am = self;
         let mut is_update = true;
@@ -125,9 +158,10 @@ pub trait ActiveModelTrait: Clone + Debug {
     }
 
     /// Delete an active model by its primary key
-    async fn delete(self, db: &DatabaseConnection) -> Result<DeleteResult, DbErr>
+    async fn delete<'a, C>(self, db: &'a C) -> Result<DeleteResult, DbErr>
     where
-        Self: ActiveModelBehavior,
+        Self: ActiveModelBehavior + 'a,
+        C: ConnectionTrait<'a>,
     {
         let am = ActiveModelBehavior::before_delete(self, db)?;
         let am_clone = am.clone();
@@ -219,23 +253,23 @@ where
         matches!(self.state, ActiveValueState::Unset)
     }
 
-    pub fn take(&mut self) -> V {
+    pub fn take(&mut self) -> Option<V> {
         self.state = ActiveValueState::Unset;
-        self.value.take().unwrap()
+        self.value.take()
     }
 
     pub fn unwrap(self) -> V {
         self.value.unwrap()
     }
 
-    pub fn into_value(self) -> Value {
-        self.value.unwrap().into()
+    pub fn into_value(self) -> Option<Value> {
+        self.value.map(Into::into)
     }
 
     pub fn into_wrapped_value(self) -> ActiveValue<Value> {
         match self.state {
-            ActiveValueState::Set => ActiveValue::set(self.into_value()),
-            ActiveValueState::Unchanged => ActiveValue::unchanged(self.into_value()),
+            ActiveValueState::Set => ActiveValue::set(self.into_value().unwrap()),
+            ActiveValueState::Unchanged => ActiveValue::unchanged(self.into_value().unwrap()),
             ActiveValueState::Unset => ActiveValue::unset(),
         }
     }