+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.6 - 2021-10-09
+
+- [[#224]] [sea-orm-cli] Date & Time column type mapping
+- Escape rust keywords with `r#` raw identifier
+
+[#224]: https://github.com/SeaQL/sea-orm/pull/224
+
+## 0.2.5 - 2021-10-06
+
+- [[#227]] Resolve "Inserting actual none value of Option<Date> results in panic"
+- [[#219]] [sea-orm-cli] Add `--tables` option
+- [[#189]] Add `debug_query` and `debug_query_stmt` macro
+
+[#227]: https://github.com/SeaQL/sea-orm/issues/227
+[#219]: https://github.com/SeaQL/sea-orm/pull/219
+[#189]: https://github.com/SeaQL/sea-orm/pull/189
+
 ## 0.2.4 - 2021-10-01
 
 - [[#186]] [sea-orm-cli] Foreign key handling